Faffing around with variable names
I know I'm being silly, renaming 'msg' to 'monosodium_glutamate', but when
I glance at 'msg', I honestly can't think of anything else. Let us rename for
shits and giggles.

This entire project and everything in it is a means to an end. That end is
having a giggle and enjoying life and being happy. Why not cut out the
middle-woman?