Refactored ComfyBlogPostMethods location to a concern
Still not totally sure app/models/concerns is the right location.
But probably better than config/initializers. Made this less
non-Railsy. I faffed around a bit with trying to get Rails to
autoload this in app/models, so as to avoid the manual
initial config class_eval call, but couldn't. This'll do.