Replaced hardcoded nav blog category links with a find_each
It is still a mystery to me why the query I have written completely
disregards my ORDER query. Damn inconvenient and damn puzzling.
May be one of those little finicky bugs that is obvious in retrospect.
All I can say is, retrospection sure hasn't happened yet.

But it's not a show-stopper. Moving on.